requirements.txt
streamlit==1.36.0
pandas==2.2.2
pyarrow==16.1.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
openai==1.68.2
